+++ CHANGES.rst
@@ -6,7 +6,7 @@ Change log for risclog.logging
 1.3.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Unify default log level and set it from DEBUG to WARNING.
 
 
 1.3.0 (2025-02-06)